requirements.txt
django==3.2.16
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
django-environ==0.9.0
psycopg2==2.9.5